+++ Sources/MapGL/Model/ISelectable.swift
@@ -1,0 +1,8 @@
+public protocol ISelectable {
+
+	/// Select object on map
+	func select()
+	/// Deselect object on map
+	func deselect()
+
+}