+++ Tests/MapGLTests/JSExecutor.swift
@@ -11,3 +11,4 @@ class JSExecutor : JSExecutorProtocol {
 	}
 
 }
+